Refactorized react and fuxed controller

+++ src/Http/Controllers/CronjobsController.php
@@ -7,6 +7,7 @@ use App\Http\Controllers\Controller;
 
 use Cron\CronExpression;
 use DateTime;
+use Exception;
 use Sefirosweb\LaravelCronjobs\Http\Models\Cronjob;
 use Sefirosweb\LaravelCronjobs\Http\Requests\CronExpressionRequest;
 use Sefirosweb\LaravelCronjobs\Http\Requests\CronjobRequest;
@@ -17,17 +18,28 @@ class CronjobsController extends Controller
     /**
      * Display a listing of the resource.
      *
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
-    public function get()
+    public function get(Request $request)
     {
-        return response()->json(['success' => true, 'data' => Cronjob::all()]);
+        $query = Cronjob::query();
+
+        if ($request->status === 'all') {
+            $query->withTrashed();
+        } else  if ($request->status === 'deleted') {
+            $query->onlyTrashed();
+        }
+
+        $data = $query->get();
+
+        return response()->json(['success' => true, 'data' => $data]);
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\CronjobRequest $request
+     * @param  Sefirosweb\LaravelCronjobs\Http\Requests\CronjobRequest $request
      * @return \Illuminate\Http\Response
      */
     public function store(CronjobRequest $request)
@@ -39,36 +51,61 @@ class CronjobsController extends Controller
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\CronjobRequest $request
-     * @param  Sefirosweb\LaravelCronjobs\Http\Models\AccessList $accessList
+     * @param  Sefirosweb\LaravelCronjobs\Http\Requests\CronjobRequest $request
      * @return \Illuminate\Http\Response
      */
-    public function update(CronjobRequest $request, Cronjob $cronjob)
+    public function update(CronjobRequest $request)
     {
-        $accessList = Cronjob::findOrFail($request->id);
+        $accessList = Cronjob::findOrFail($request->cronjob_id);
         $accessList->update($request->all());
         return response()->json(['success' => true]);
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Enable & Disable Cronjob
      *
-     * @param  \Sefirosweb\LaravelCronjobs\Http\Models\AccessList $accessList
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function destroy(Request $request)
     {
-        $cronjob = Cronjob::findOrFail($request->id);
-        $cronjob->delete();
+        $cronjob = Cronjob::withTrashed()->findOrFail($request->cronjob_id);
+        if (!$cronjob->deleted_at) {
+            $cronjob->next_run_at = null;
+            $cronjob->save();
+            $cronjob->delete();
+        } else {
+            try {
+                $nextRunAt = $this->calculate_next_run($cronjob->cron_expression);
+            } catch (Exception $e) {
+                $nextRunAt = null;
+            }
+            $cronjob->next_run_at = $nextRunAt;
+            $cronjob->save();
+            $cronjob->restore();
+        }
         return response()->json(['success' => true]);
     }
 
+    /**
+     * Enable & Disable Cronjob
+     *
+     * @param  Sefirosweb\LaravelCronjobs\Http\Requests\CronExpressionRequest $request
+     * @return \Illuminate\Http\Response
+     */
     public function preview_job(CronExpressionRequest $request)
     {
         $nextRun = $this->calculate_next_run($request->inputCroExpresion, 40);
         return response()->json(['success' => true, 'data' => $nextRun]);
     }
 
+    /**
+     * Calculate the time date to next run based on $cronExpresion
+     *
+     * @param String $cronExpresion
+     * @param Int $manyRuns
+     * @return String
+     */
     private function calculate_next_run($cronExpresion, $manyRuns = 1)
     {
         $cron = new CronExpression($cronExpresion);
@@ -84,6 +121,12 @@ class CronjobsController extends Controller
         return $runs;
     }
 
+    /**
+     * Edit cronjob rule
+     *
+     * @param  Sefirosweb\LaravelCronjobs\Http\Requests\CronExpressionRequest $request
+     * @return \Illuminate\Http\Response
+     */
     public function edit_cron_timer(CronExpressionRequest $request)
     {
         $nextRun = $this->calculate_next_run($request->inputCroExpresion);
@@ -110,7 +153,7 @@ class CronjobsController extends Controller
     {
         $dateNow = new DateTime();
 
-        $jobs = Cronjob::where('next_run_at', '<=', $dateNow)->get();
+        $jobs = Cronjob::whereNot()->where('next_run_at', '<=', $dateNow)->get();
         logger("Executing kernel cronjob, current pending jobs: " . $jobs->count());
 
         $jobs->each(function ($job) {
